requirements/base.txt
appdirs==1.4.3
beautifulsoup4==4.6.0
certifi==2018.4.16
chardet==3.0.4
Django==1.11.13
django-environ==0.4.1
django-modelcluster==4.1
django-taggit==0.22.2
django-treebeard==4.3
djangorestframework==3.8.2
draftjs-exporter==2.1.1
html5lib==1.0.1
idna==2.6
l18n==2016.6.4
olefile==0.44
packaging==16.8
Pillow==5.1.0
psycopg2==2.6.2
Pygments==2.1.3
pyparsing==2.2.0
pytz==2018.4
requests==2.18.4
six==1.11.0
taggable==0.2.0
Unidecode==0.4.21
urllib3==1.22
wagtail==2.1
wagtail-pg-search-backend==1.3.1
wagtailmedia==0.2.0
wagtailmenus==2.9
webencodings==0.5.1
Willow==1.1
